requirements.txt
pygame
google-genai
python-dotenv
google-generativeai
python-dotenv